backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
openai>=1.12.0
httpx>=0.27.0
python-multipart==0.0.6
pyyaml>=6.0
